# Cache the results of SwiftASTContext::GetCompileUnitImports()