Every time a scratch context is requested, a call to
GetCompileUnitImports() is made, which is both expensive and
redundant. This patch adds a cache that keeps track of compile units
for which the imports were already performed. This results in a
measurable performance improvement on the LLDB test suite.

rdar://75381959

The results are very noisy, so please take this with a grain of salt.
Before:
```
Tests Times:
--------------------------------------------------------------------------
[   Range   ] :: [               Percentage               ] :: [ Count ]
--------------------------------------------------------------------------
[150s,160s) :: [                                        ] :: [  1/355]
[140s,150s) :: [                                        ] :: [  0/355]
[130s,140s) :: [                                        ] :: [  0/355]
[120s,130s) :: [                                        ] :: [  0/355]
[110s,120s) :: [                                        ] :: [  0/355]
[100s,110s) :: [                                        ] :: [  0/355]
[ 90s,100s) :: [                                        ] :: [  0/355]
[ 80s, 90s) :: [                                        ] :: [  0/355]
[ 70s, 80s) :: [                                        ] :: [  1/355]
[ 60s, 70s) :: [                                        ] :: [  0/355]
[ 50s, 60s) :: [                                        ] :: [  7/355]
[ 40s, 50s) :: [                                        ] :: [  2/355]
[ 30s, 40s) :: [*                                       ] :: [ 11/355]
[ 20s, 30s) :: [****                                    ] :: [ 40/355]
[ 10s, 20s) :: [*****                                   ] :: [ 51/355]
[  0s, 10s) :: [***************************             ] :: [242/355]
--------------------------------------------------------------------------
```

After:

```
Tests Times:
--------------------------------------------------------------------------
[   Range   ] :: [               Percentage               ] :: [ Count ]
--------------------------------------------------------------------------
[140s,150s) :: [                                        ] :: [  1/355]
[130s,140s) :: [                                        ] :: [  0/355]
[120s,130s) :: [                                        ] :: [  0/355]
[110s,120s) :: [                                        ] :: [  0/355]
[100s,110s) :: [                                        ] :: [  0/355]
[ 90s,100s) :: [                                        ] :: [  0/355]
[ 80s, 90s) :: [                                        ] :: [  0/355]
[ 70s, 80s) :: [                                        ] :: [  1/355]
[ 60s, 70s) :: [                                        ] :: [  0/355]
[ 50s, 60s) :: [                                        ] :: [  3/355]
[ 40s, 50s) :: [                                        ] :: [  5/355]
[ 30s, 40s) :: [                                        ] :: [  7/355]
[ 20s, 30s) :: [****                                    ] :: [ 42/355]
[ 10s, 20s) :: [*****                                   ] :: [ 50/355]
[  0s, 10s) :: [***************************             ] :: [246/355]
--------------------------------------------------------------------------
```